requirements.txt
#django==4.2.20
#django-autocomplete-light==3.9.7
#django-braces==1.16.0
#django-crispy-forms==2.3
#django-dbbackup==4.2.1
#django-debug-toolbar==4.3.0
#django-filter==24.3
#django-localflavor==4.0
#crispy-bootstrap5==2025.4
#psycopg2==2.9.10
#gunicorn

#Meu projeto
asgiref==3.8.1                                                                                                                                                         
astroid==3.3.9                                                                                                                                                         
autopep8==2.3.2
colorama==0.4.6
crispy-bootstrap4>=2022.1
dill==0.4.0
distlib==0.3.9
Django>=5.2
django-braces==1.17.0
django-crispy-forms>=1.14.0
django-extensions==4.1
filelock==3.18.0
gunicorn>=21.2.0
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
packaging==24.2
pbr==6.1.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary>=2.9.0
pycodestyle==2.13.0
pylint==3.3.6
pytest==8.3.5
python-dotenv>=0.19.0
setuptools==78.1.0
sqlparse==0.5.3
stevedore==5.4.1
tomlkit==0.13.2
tzdata==2025.2
virtualenv==20.29.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
virtualenvwrapper-win==1.2.7
django-debug-toolbar>=4.4
django-filter>=24.2
